Add files via upload
new weatherscene2 was added where precipitation changes state as it falls down the screen. WeatherScene should be the same, but MenuScene is loading WeatherScene2 as of right now. It seems that adding the extra checks in precipFalls() made the weatherScene slow down to an unrunnable point.

In precipFalls(), whereIsTheWeather is called to observe what the state of the precipitation should be and it is checked against the precipitations current physicsbody.categorytype. If the two are different, the new precipitation is created at the location of the old one. 


Other notable changes are the addition of the method isThereWeather(cloud: SKSpriteNode), which returns a boolean of whether the given SKSpriteNode is over the mountains or not. (returns false if node is over lake).
The method: whereIsTheWeather(cloud: SKSpriteNode) now returns the given precipitation type based on the y coord of the given node.

thunderAndLightning was placed in precipFalls() in an attempt to reduce the system requirements (with no avail).