Add files via upload
i've introduced an apple tree using the cloud sprite. It currently makes an apple on touch that can be dragged after created. An optional additional feature would be having an apple be created and dragged away as a result of trying to move the appletree. 

A known bug is that when the horse eats a bunch of apples, multiple horses are created. Rather than trying to fix this, I suggest we first try to alter our code to have the horse's fart vary based on the number of apples eaten. 
If each time the horse eats an apple, the timer is discarded and a new timer is created preparing for the next tier of fart, then the horse will eventually fart only one, and the chunk of code being executed when the final fart timer finishes seems to be the code that is creating multiple horses.

In other news, the weather on weatherScene2 is working again. I added the 
                                                   if shouldItRain % 10 == 0 {~~~~~~~~~~}
      clause to the chunk of precip falls that checks sound and checks if any precip should be changed to its new zone. It seems that there were just too many operations happening in update. This makes me think that we ought to not have so many things in update. We should at least try to loop through the precipitation array as little as possible in update, as that array can grow to be 200 nodes at a time. 

I think the uniformity of the horizontal zones looks somewhat unrealistic. I think it would look much better if the snow stopped at the top of the high mountain. We could increment the alpha value of a sprite that we put there called mountainTop to make it appear as if the snow slowly gathers and creates a snowy peak. The sun could then increment the lake's water level and decrement that alpha value (or blend with a brownish green, it might look more realistic) to simulate snow melting.