Add files via upload
    static let Hail      : UInt32 = 0b1001     //9
    static let Snow      : UInt32 = 0b1010     //10
    static let Ground    : UInt32 = 0b1011     //11
    static let SnowCap   : UInt32 = 0b1100     //12

I added some more physicsCategories, but it just occurred to me you probably did too, so we'll have to change something. The hail and rain are still not appearing sometimes. The only bug fix i can think of that i havent tried it checking their position when they are not generated. 
The bug is especially odd because you can use the same cloud on the same instance of running the app and the hail and rain will work only sometimes. The sounds work which means the chunk of code in precipFalls where the precip switches types as it falls is being accessed, but for some reason they're not being added to the background. The hail sound is commented out right now because I didn't have it in my Assets folder. 
The nodes on the screen are also reflecting that the rain and hail are, in fact, on the screen, just not, y'know, on the screen.