Add files via upload
Changes mostly included formatting and resizing the animal systems to be kid friendly sizes. Another mentionable change was the introduction of a mouth for each of the animals. The sprite used is a smaller version of the sun as it had a good shade for both of the animals heads. The mouths are both being added at a depth of -2 so they are behind the background and never appear. The purpose of the mouths is so that the animals only eat using their heads. The previous system (a collision with the animal as well as the food being within bounds that defined the animal's head) was more costly and somewhat finicky.

The apples aren't appearing on the tree in an attractive manner. I fussed with the numbers in the apple generation code, but I didn't get it perfect. 

The code could still needs to be groomed for stylistic consistency. 

I reduced the mouth size several times but the animals are still eating food a little bit far outside of their reach. This could be tinkered with to make it look better.

The animal buttons were adjusted to represent their entire animal-foodSupply, food systems. They should add and remove their whole systems now. Note that the screen gets rather cluttered if both systems are present.

Feathers were added. They're represented in an array; they are falling via feathersFall(array: feathers) in update. They appear whenever the lion eats a chicken. 
I also added a squawking sound whenever a chicken is generated or a chicken is eaten. We could introduce a second chicken sound for either of the previously mentioned situations as it may be confusing to have the same sound for birth and death, but isn't that how life goes?

I've included a bunch of files for weatherScene and BoxScene alike. My apologies for the few that are named IMG_238282 or what have you. You'll have to check what they are in the initializing sections.